+++ NEWS.md
@@ -1,3 +1,10 @@
+# HEAD
+
+### Bug fixes
+
+* Fix `validate_inclusion_of` + `in_array` when used against a date or datetime
+  attribute so that it does not raise a CouldNotSetAttributeError.
+
 # 3.0.0
 
 ### Backward-incompatible changes